# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SecureShare API",
        "description": "Secure Document Collaboration Platform API with RESTful endpoints for project management, document sharing, and team collaboration.",
        "contact": {
            "email": "support@secureshare.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/projects/{projectId}/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "List project documents",
                "description": "List project documents",
                "operationId": "71cfc9bae52bd270931e3c8a27731539",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Documents list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}